addresses PR feedback
