+++ cli/cmd/testdata/deployment_test_file.yml
@@ -34,4 +34,3 @@ spec:
           name: http
         resources: {}
 status: {}
----
